feat: Add spreadsheet URL setting
Introduces a new `spreadsheetUrl` field to `SystemSettings` to store the URL for external spreadsheets. This allows for better integration with external data sources.

The `OnCall` component and `App` component have been updated to pass and utilize this new setting. The `Settings` component has been enhanced to manage and save this URL.