feat: Remove manageableProfiles from RoleConfig
The `manageableProfiles` field in `RoleConfig` and its related logic have been removed. This feature was intended to restrict which user profiles a role could manage, but this functionality is no longer required and has been deprecated. The `UserProfile` type has also been updated to be a string to align with this change.