feat: Pass availableBranches to Collaborators component
This change introduces the `availableBranches` prop to the `Collaborators` component. It is also passed down from `App.tsx` to ensure that the component has access to the list of branches the current user is allowed to interact with.

This allows for auto-selection of the user's branch when creating a new collaborator if they are restricted to a single branch.